producer/requirements.txt
confluent-kafka
fastapi
kafka-python
pandas
uvicorn
Authlib
python-multipart
pyjwt